# backend/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Six Seven API",
    "version": "1.0.0",
    "description": "Backend API for Six Seven - Talk the Alpha, Walk the Alpha",
    "contact": {
      "name": "Six Seven Support",
      "email": "support@sixseven.app"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/chat": {
      "post": {
        "summary": "Send a chat message",
        "description": "Process a chat message with optional image and generate AI response in Gen Alpha slang",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "The text message to process"
                  },
                  "image": {
                    "type": "string",
                    "description": "Base64 encoded image for OCR processing"
                  },
                  "model": {
                    "type": "string",
                    "enum": [
                      "1x",
                      "2x",
                      "3x",
                      "4x"
                    ],
                    "default": "1x",
                    "description": "The alpha level model to use"
                  },
                  "conversationId": {
                    "type": "string",
                    "description": "Existing conversation ID to continue the chat"
                  }
                },
                "oneOf": [
                  {
                    "required": [
                      "text"
                    ]
                  },
                  {
                    "required": [
                      "image"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string",
                      "description": "The AI generated response"
                    },
                    "credits": {
                      "type": "integer",
                      "description": "Remaining daily credits"
                    },
                    "maxCredits": {
                      "type": "integer",
                      "description": "Maximum daily credits"
                    },
                    "conversationId": {
                      "type": "string",
                      "description": "The conversation ID"
                    },
                    "messageId": {
                      "type": "string",
                      "description": "The message ID"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing text or image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Daily limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "credits": {
                      "type": "integer"
                    },
                    "maxCredits": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User or conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stripe/sync-subscription": {
      "post": {
        "summary": "Sync subscription status",
        "description": "Sync the user's subscription status with Stripe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription synced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "subscriptionEndsAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true,
                      "description": "Subscription end date"
                    },
                    "status": {
                      "type": "string",
                      "description": "Subscription status from Stripe"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No subscription found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Failed to sync subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}